call api in componentDidMount to get data every time  on refresh
