+++ src/containers/Pages/TutorialPage.js
@@ -2,15 +2,20 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
 import Footer from "../../components/Footer";
+import { getTutorialById } from "../../store/actions/Actions";
 import "./CoursePage.scss";
 
 class TutorialPage extends Component {
+  componentDidMount() {
+    this.props.getTutorials(this.props.history.location.state);
+  }
+
   render() {
     return (
       <div className="main-course-container">
         <div className="header">
           <h1>
-            {this.props.tutorialById[0] && this.props.tutorialById[0].name}
+            {this.props.tutorialById ? this.props.tutorialById[0].name : null}
           </h1>
         </div>
         <div className="course-content-container">
@@ -18,8 +23,9 @@ class TutorialPage extends Component {
             <div className="card shadow-lg">
               <div className="card-body">
                 <p>
-                  {this.props.tutorialById[0] &&
-                    this.props.tutorialById[0].content}
+                  {this.props.tutorialById
+                    ? this.props.tutorialById[0].content
+                    : null}
                 </p>
               </div>
             </div>
@@ -35,4 +41,16 @@ const mapStateToProps = state => {
     tutorialById: state.getTutorialDataById.tutorialById
   };
 };
-export default withRouter(connect(mapStateToProps)(TutorialPage));
+const mapDispatchToProps = dispatch => {
+  return {
+    getTutorials: id => {
+      dispatch(getTutorialById(id));
+    }
+  };
+};
+export default withRouter(
+  connect(
+    mapStateToProps,
+    mapDispatchToProps
+  )(TutorialPage)
+);
