+++ src/store/reducers/Reducer.js
@@ -23,6 +23,8 @@ const userData = (state = initialState, action) => {
   return newState;
 };
 const courseData = (state = initialState, action) => {
+  debugger;
+
   const newState = { ...state };
   if (action.type === types.GET_COURSE_DATA) {
     return {
@@ -35,6 +37,7 @@ const courseData = (state = initialState, action) => {
 };
 
 const tutorialData = (state = initialState, action) => {
+  debugger;
   const newState = { ...state };
   if (action.type === types.GET_TUTORIAL_DATA) {
     return {
@@ -46,6 +49,7 @@ const tutorialData = (state = initialState, action) => {
 };
 
 const quizData = (state = initialState, action) => {
+  debugger;
   const newState = { ...state };
   if (action.type === types.GET_QUIZ_DATA) {
     return {
@@ -57,6 +61,7 @@ const quizData = (state = initialState, action) => {
 };
 
 const getTutorialDataById = (state = initialState, action) => {
+  debugger;
   const newState = { ...state };
   if (action.type === types.GET_TUTORIAL_BY_ID) {
     return {
